backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlalchemy[asyncio]==2.0.30
asyncpg==0.29.0
alembic==1.13.1
pydantic==2.7.1
pydantic-settings==2.3.0
redis[hiredis]==5.0.4
twilio==9.0.4
livekit>=1.1.2,<2.0.0
livekit-agents>=1.0.11,<2.0.0
livekit-plugins-deepgram==0.6.5
livekit-plugins-openai==0.8.1
livekit-plugins-elevenlabs==0.7.2
livekit-plugins-silero>=1.0.11,<2.0.0
livekit-plugins-noise-cancellation>=0.2.3,<1.0.0
openai>=1.35.0,<2.0.0
httpx==0.27.0
python-dotenv==1.0.1
stripe==9.9.0
sentry-sdk[fastapi]==2.5.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
